# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Koine Gateway API
  version: 1.0.0
  description: HTTP gateway service for Claude Code CLI. Provides REST endpoints for text generation, structured object generation, and streaming responses.
  license:
    name: AGPL-3.0-only OR LicenseRef-Commercial
    url: https://github.com/pattern-zones-co/koine/blob/main/LICENSE
  contact:
    name: Pattern Zones Co
    url: https://github.com/pattern-zones-co/koine
servers:
  - url: http://localhost:3100
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. Generate with: openssl rand -hex 32"
...
paths:
...
  /stream:
    post:
      summary: Stream text response via SSE
      description: "Streams Claude CLI output using Server-Sent Events (SSE). Provides real-time streaming of Claude's response with session management and usage tracking. Event types: session (session ID), text (streaming content), result (final usage stats), error (errors), done (stream complete)."
      tags:
        - Streaming
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                system:
                  type: string
                prompt:
                  type: string
                sessionId:
                  type: string
                model:
                  type: string
              required:
                - prompt
      responses:
        "200":
          description: SSE stream with real-time Claude CLI output
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
        "401":
          description: Missing authorization header
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
  /stream-object:
    post:
      summary: Stream structured JSON objects via SSE
      description: "Streams partial JSON objects as they're generated using Server-Sent Events (SSE). Provides real-time streaming of structured data with partial object parsing. Event types: session (session ID), partial-object (partial JSON with parsed object), object (final validated object), result (final usage stats), error (errors), done (stream complete)."
      tags:
        - Streaming
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                system:
                  type: string
                prompt:
                  type: string
                sessionId:
                  type: string
                model:
                  type: string
                schema:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: *a1
                    properties:
                      type: object
                      additionalProperties: {}
                    items: {}
                    $ref:
                      type: string
                    allOf:
                      type: array
                      items: {}
                    anyOf:
                      type: array
                      items: {}
                    oneOf:
                      type: array
                      items: {}
                    enum:
                      type: array
                      items: {}
                    const: {}
              required:
                - prompt
                - schema
      responses:
        "200":
          description: SSE stream with real-time partial JSON objects
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
        "401":
          description: Missing authorization header
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
...